Move MaterialOutputModel back
